Functions

Base.:* — Method

*(a::LSpec, b::RSpec)

results in a luminance spectrum LSpec due to an illuminant spectrum reflected by a reflectance spectrum.

Result: LSpec

Base.:* — Method

*(a::LSpec, a::TSpec)

results in a luminance spectrum LSpec due to an illuminant filtered by a transmission spectrum.

Base.:* — Method

(a::Array{Real,2}, c::ConeFund)

Convert cone fundamentals to CMFs.

Base.:* — Method

*(a::RSpec, b::RSpec)

Interreflection between two reflective surfaces

Example:

julia> macbeth(:Macbeth_chart1, :red) * macbeth(:Macbeth_chart1, :orange)
RSpec(Real[380.0, 390.0, 400.0, 410.0, 420.0, 430.0, 440.0, 450.0, 460.0, 470.0  …  640.0, 650.0, 660.0, 670.0, 680.0, 690.0, 700.0, 710.0, 720.0, 730.0], Real[0.002597, 0.002544, 0.0024909999999999997, 0.0024909999999999997, 0.0024909999999999997, 0.002538, 0.002538, 0.002484, 0.0024749999999999998, 0.00252  …  0.376929, 0.394878, 0.40763199999999994, 0.42066499999999996, 0.43686499999999995, 0.4553, 0.472696, 0.48322000000000004, 0.486576, 0.49293200000000004])

Base.:* — Method

*(a::TSpec, b::TSpec)

Stacking of two filters.

Results in a transmission spectrum TSpec due to the combination of two transmissive spectra using Bouguer’s/Beer’s Law.

Base.:+ — Function

+(a::RSpec, b::RSpec, ratio=0.5)

Linear combination of two reflectance spectra at a given ratio r * a / (1 - r) * b.

Warning: This is not a physically realistic colorant mixing method!

Examples

Base.:+ — Method

+(a::LSpec, b::LSpec)

Combination of two luminance spectra. Additive mixture.

Examples:

julia> D_series_illuminant(6504) + D_series_illuminant(4010)
LSpec(Real[390.0, 391.0, 392.0, 393.0, 394.0, 395.0, 396.0, 397.0, 398.0, 399.0  …  821.0,
822.0, 823.0, 824.0, 825.0, 826.0, 827.0, 828.0, 829.0, 830.0], Real[71.28837977644169,
75.09766988566693, 78.90695999489216, 82.7162501041174, 86.52554021334264, 90.33483032256788,
94.14412043179311, 97.95341054101836, 101.7627006502436, 105.57199075946883  … 148.07280275201228,
148.84652207565884, 149.6202413993054, 150.39396072295196, 151.16768004659852, 151.94139937024508,
152.7151186938916, 153.48883801753817, 154.26255734118473, 155.0362766648313])

SpectralVis.:× — Method

×(l::LSpec, cmf::CMatch)

Create white point of a light source in XYZ color space.

Examples:

SpectralVis.:× — Method

×(l::LSpec, r::RSpec, cmf::CMatch)

Returns XYZ color of an illuminant spectrum reflected off a surface.

Examples:

SpectralVis.:× — Method

×(l::LSpec, t::TSpec, cmf::CMatch)

Returns the XYZ color of an illuminant transmitted through a filter.

Examples:

SpectralVis.D40_illuminant — Method

D40_illuminant()

Result: ::LSpec of D40 illuminant, with a CCT of 4003 K

SpectralVis.D50_illuminant — Method

D50_illuminant()

Result: ::LSpec of D50 illuminant, with a CCT of 5003 K

SpectralVis.D55_illuminant — Method

D55_illuminant()

Result: ::LSpec of D55 illuminant, with a CCT of 5503 K

SpectralVis.D65_illuminant — Method

D65_illuminant()

Result: ::LSpec of D55 illuminant, with a CCT of 6504 K

SpectralVis.D75_illuminant — Method

D75_illuminant()

Result: ::LSpec of D75 illuminant, with a CCT of 7504 K

SpectralVis.D93_illuminant — Method

D93_illuminant()

Result: ::LSpec of D93 illuminant, with a CCT of 9304 K

SpectralVis.D_series_generator — Method

D_series_generator(T::Real)

The relative spectral power distribution SPD(λ) of a D series illuminant at CCT T, from 300 to 830 nm in 5 nm steps. By definition the output spectra are normalized in a way that their luminance is always 100.0 at a wavelength of 560 nm.

Allowed CCT range: 4000 K ≤ CCT ≤ 25000 K

SpectralVis.D_series_illuminant — Method

D_series_illuminant(T::Real)

produces CIE Daylight spectrum according to desired CCT.

result: ::LSpec

SpectralVis.D_series_illuminant — Method

D_series_illuminant(env::SpecEnv, T::Real)

produces CIE Daylight spectrum according to desired CCT.

result: ::LSpec

SpectralVis.D_series_proper_illuminant — Method

D_series_proper_illuminant(T::Real)

Generates a D-series illuminant with the proper CCT. For example, a D65 illuminant has a proper CCT of 6504 K. CCT according to the International Practical Temperature Scale from 1968.

SpectralVis.D_series_whitepoint — Method

D_series_whitepoint(T::Real)

produces the white point according to the CIE Daylight Series.

result: ::XYZ

SpectralVis.D_series_whitepoint — Method

D_series_whitepoint(env::SpecEnv,T::Real)

produces the white point according to the CIE Daylight Series.

result: ::XYZ

SpectralVis.F_series_illuminant — Method

F_series_illuminant(n::Int)

Return F series illuminant Fn for n from 1 to 12.

Examples

julia> F_series_illuminant(1)
LSpec(Real[390.0, 391.0, 392.0, 393.0, 394.0, 395.0, 396.0, 397.0, 398.0, 399.0  …  821.0, 822.0, 823.0, 824.0, 825.0, 826.0, 827.0, 828.0, 829.0, 830.0], Real[2.94, 3.046, 3.152, 3.258, 3.3640000000000003, 3.47, 3.81, 4.15, 4.49, 4.83  …  0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0])

julia> F_series_illuminant(13)
ERROR: DomainError with 13:
F series illuminant outside the range 1 ≤ n ≤ 12.

SpectralVis.F_series_illuminant — Method

F_series_illuminant(type::Symbol)

Spectral power distribution of representative fluorescent lamps from 380 to 780 nm in 5 nm steps.

Available types:

:f1 ... :f12

F2, F7, and F11 should be preferred over the other distributions as long as the choice in each of these three groups is not important or critical. The three recommendend illuminants are also marked in the CRI lists below.

For wavelength outside the 380 nm ... 780 nm range the SPDs should be set to zero.

Result: ::LSpec

These are not CIE Standard Illuminants. They were compiled by the CIE as a representative collection of practical fluorescent lamps. The 12 lamps in this collection can be grouped in 3 groups:

normal: two semi-broad band emmisions of antimony and manganese activations in calcium halo-phosphate phosphor.

CRI table

F1: 76

F2: 64

F3: 57

F4: 51

F5: 72

F6: 59

broad-band: enhanced color rendering properties in comparison to the normal group, mostly by using multiple phosphors. Flatter spectral distributions with a wider spectral range.

CRI table

F7: 90

F8: 95

F9: 90

three-band: three narrow-band emissions in the red, green, and blue spectral regions. The narrow-band emissions are produced by ternary compositions of rare-earth phosphors.

CRI table

F10: 81

F11: 83

F12: 83

Examples

julia> F_series_illuminant(:f11)
LSpec(Real[390.0, 391.0, 392.0, 393.0, 394.0, 395.0, 396.0, 397.0, 398.0, 399.0  …  821.0, 822.0, 823.0, 824.0, 825.0, 826.0, 827.0, 828.0, 829.0, 830.0], Real[0.46, 0.442, 0.424, 0.406, 0.388, 0.37, 0.554, 0.738, 0.922, 1.1059999999999999  …  0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0])

julia> F_series_illuminant(:f0)
ERROR: DomainError with f0:
F-series illuminant type does not exist.

SpectralVis.adapt_spec — Function

`adapt_spec(spec::Spectrum, specenv::SpecEnvironment = SPECENV, extraptype::Symbol = specenv.ex, interporder...)`

adapts spectra according to the spectral environment.

Arguments for extraptype:

:zero, sets SPD outside the data range to zero.

:boundary, sets SPD outside the data range to boundary values.

:linear, linear extrapolation of SPD.

:quadratic, quadratic extrapolation of SPD.

Arguments for interporder...:

:linear for linear splines

:quadratic for quadratic splines

:cubic, :natural for natural cubic splines (zero curvature at boundaries).

:cubic, :periodic for periodic cubic splines (both endpoints must have same value).

:cubic, :clamped, slope_start, slope_end for clamped cubic splines with defined slopes at boundaries.

:cubic, :constrained for cubic constrained splines.

Examples

julia> adapt_spec(agfait(2), SPECENV, :linear, :cubic, :natural)
RSpec(Real[390.0, 391.0, 392.0, 393.0, 394.0, 395.0, 396.0, 397.0, 398.0, 399.0  …  821.0, 822.0, 823.0, 824.0, 825.0, 826.0, 827.0, 828.0, 829.0, 830.0], Real[0.06193665145465862, 0.06273298630919276, 0.06352932116372689, 0.06432565601826104, 0.06512199087279517, 0.06591832572732931, 0.06671466058186346, 0.06751099543639759, 0.06830733029093172, 0.06910366514546587  …  0.44181513985324095, 0.444222702992524, 0.4466302661318069, 0.4490378292710899, 0.45144539241037296, 0.4538529555496559, 0.4562605186889389, 0.45866808182822183, 0.46107564496750486, 0.4634832081067878])

julia> adapt_spec(D65_illuminant(), SPECENV, :zero, :cubic, :clamped, 0.0, 0.0)
LSpec(Real[390.0, 391.0, 392.0, 393.0, 394.0, 395.0, 396.0, 397.0, 398.0, 399.0  …  821.0, 822.0, 823.0, 824.0, 825.0, 826.0, 827.0, 828.0, 829.0, 830.0], Real[54.666490682108055, 57.4776549221071, 60.288819162106144, 63.099983402105195, 65.91114764210424, 68.72231188210328, 71.53347612210233, 74.34464036210137, 77.15580460210042, 79.96696884209946  …  57.72152419756936, 58.00867916043931, 58.29583412330925, 58.5829890861792, 58.870144049049145, 59.15729901191909, 59.44445397478904, 59.73160893765898, 60.01876390052893, 0.0])

SpectralVis.adapted_color — Method

adaptated_color(spectrum)

SpectralVis.agfait — Method

agfait(idx::Int)

Returns the reflectance spectrum of the selected Agfait872 color patch. Spectral data are from 400 to 700 nm in 10 nm steps.

1st group: 1-264

2nd group: Neutral scale samples (265-286)

3rd group:

black sample (287)

white sample (288)

Minolta white calibration sample (289)

Examples

julia> agfait(123)
RSpec(Real[400.0, 410.0, 420.0, 430.0, 440.0, 450.0, 460.0, 470.0, 480.0, 490.0  …  610.0, 620.0, 630.0, 640.0, 650.0, 660.0, 670.0, 680.0, 690.0, 700.0], Real[0.1282, 0.3004, 0.4806, 0.5796, 0.6429, 0.6818000000000001, 0.7078, 0.7159, 0.7071999999999999, 0.6896  …  0.259, 0.2318, 0.21309999999999998, 0.2, 0.1932, 0.19210000000000002, 0.19699999999999998, 0.2095, 0.2311, 0.2594])

julia> agfait(0)
ERROR: DomainError with 0:
Color index out of range. Expecting 0 ≤ index ≤ 289.

Measurer: ela@ee.oulu.fi (University of Oulu, Electrical Eng.Dept)

SpectralVis.agfait_specs — Method

agfait_specs()

SpectralVis.blackbody_illuminant — Method

blackbody_illuminant(env::SpecEnv,T::Real)

produces blackbody spectrum according to desired blackbody temperature in K.

Examples

SpectralVis.blackbody_whitepoint — Method

blackbody_whitepoint(env::SpecEnv, T::Real)

produces the white point according to desired blackbody temperature in K.

Examples

SpectralVis.block_spec — Function

block_spec(λmin, λmax, Δλ, λs, λe, str = 1.0, mode = 0)

Regular block spectrum initialized to 0.0, with blocks λs..λe = 1.0

block_spec(λmin, λmax, Δλ, λs, λe, str = `1.0`, 1)

Inversed block spectrum initialized to 1.0, with blocks λs..λe = 0.0

SpectralVis.block_spec — Method

block_spec(λ)

Zero block spectrum with only wavelength λ set to 1.0 Fast version to create highest purity optimal colors.

Spectral limits are taken from the spectral environment.

SpectralVis.block_spec — Method

block_spec(env::SpecEnvironment, λ)

Zero block spectrum with only wavelength λ set to 1.0.

Fast version to create highest purity optimal colors. Linear interpolation for intermediate λ values.

SpectralVis.chromatic_adaptation — Function

chromatic_adaptation(sourcecolor::Color, sourceWP::Color, destWP::Color = D_series_whitepoint(6504), response::Symbol = :cat02)

Adapt source color according to given source white point and destination white point. Default destination white point is D65. Default cone response is :cat02.

Examples

julia> chromatic_adaptation(XYZ(0.1, 0.7, 0.05), D_series_whitepoint(4000))
XYZ{Float64}(0.04957926027638688,0.7081276413719826,0.09614857166385628)

julia> chromatic_adaptation(XYZ(0.1, 0.7, 0.05), D_series_whitepoint(4000), D_series_whitepoint(7500), :bradford)
XYZ{Float64}(0.0513224358507469,0.7251796437072336,0.06113763051342081)

SpectralVis.cmfunc — Method

cmfunc(conefund::ConeFund)

Returns the full multispectral color matching functions based on the cone fundamental data in cmf_tables.jl. Useful to convert (altered) cone fundamental functions back to color matching functions, e.g. for simulations of color anomalous observers with missing or shifted cone responses.

Possible conefund types:

LMS31, CIE1931 2° observer cone fundamentals

LMS64, CIE1964 10° observer cone fundamentals

LMS06_2, CIE2012 2° observer (current CIE 2° standard observer) cone fundamentals

LMS06_10, CIE2012 10° observer (current CIE 10° standard observer) cone fundamentals

Example

julia> cmfunc(conefund(:cie12_10))
CIE12_10(Real[390.0, 391.0, 392.0, 393.0, 394.0, 395.0, 396.0, 397.0, 398.0, 399.0  …  821.0, 822.0, 823.0, 824.0, 825.0, 826.0, 827.0, 828.0, 829.0, 830.0], Real[0.0001205700118214772, 0.00015063103540534432, 0.00018705601478443535, 0.0002308215774282965, 0.00028292144443875793, [...], 0.010780088746948505, 0.012937395576635658, 0.015480114358540253  …  0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0])

SpectralVis.cmfunc — Method

cmfunc(match::Symbol)

Returns the full multispectral matching functions based on the CMF data in cmf_tables.jl

Possible match symbols:

:cie31, CIE1931 2° observer

:cie64, CIE1964 10° observer

:cie31_j, CIE1931 observer with corrections by Judd

:cie31_jv, CIE1964 observer wtih correction by Judd and Vos

:cie12_2, CIE2012 2° observer (current CIE 2° standard observer)

:cie12_10, CIE2012 10° observer (current CIE 10° standard observer)

Example

julia> cmfunc(:cie12_2)
CIE12_2(Real[390.0, 391.0, 392.0, 393.0, 394.0, 395.0, 396.0, 397.0, 398.0, 399.0  …  821.0, 822.0, 823.0, 824.0, 825.0, 826.0, 827.0, 828.0, 829.0, 830.0], Real[0.003769647, 0.004532416, 0.005446553, 0.006538868, 0.007839699, 0.009382967, 0.01120608, 0.01334965, 0.0158569, 0.01877286  …  2.986206e-6, 2.814999e-6, 2.653663e-6, 2.501725e-6, 2.358723e-6, 2.224206e-6, 2.097737e-6, 1.978894e-6, 1.867268e-6, 1.762465e-6], Real[0.0004146161, 0.0005028333, 0.0006084991, 0.0007344436, 0.0008837389, 0.001059646, 0.001265532, 0.001504753, 0.001780493, 0.002095572  …  1.190771e-6, 1.123031e-6, 1.059151e-6, 9.989507e-7, 9.422514e-7, 8.888804e-7, 8.38669e-7, 7.914539e-7, 7.47077e-7, 7.05386e-7], Real[0.0184726, 0.02221101, 0.02669819, 0.03206937, 0.03847832, 0.04609784, 0.05511953, 0.06575257, 0.07822113, 0.09276013  …  0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0])

SpectralVis.cohen_pair — Function

cohen_pair(illum::LSpec, rad::Spectrum, match::ConeFund = conefund(SPECENV.cmf), type::Symbol = :standard)

Returns a pair of Cohen’s fundamental metamer and metameric black according to a light stimulus.

Examples

SpectralVis.col — Method

col(type,lum::LSpec,whitepoint::Color3,colmatch,dl)

This function creates the color loci of the theoretically maximum color gamut under a given illuminant.

type: color space

lum: illuminant

whitepoint: XYZ whitepoint

colmatch: color matching function

dl: delta lambda

output: a,b,c (locus vectors in given color space)

    x,y,Y for Yxy

    X,Z,Y for XYZ

    a,b,L Lab, DIN99, DIN99d, DIN99o

    u,v,L for Luv

SpectralVis.colorvisiondeficiency — Function

colorvisiondeficiency(model::Symbol, deftype::Symbol, c::Spectrum, illum::LSpec = D65_illuminant() , shift::Real = 0.0)

color vision deficiency simulation using Jozef Cohen’s matrix-R method.

SpectralVis.colorvisiondeficiency — Function

colorvisiondeficiency(deftype::Symbol, Δλ::Real = 0.0)

Returns the cone fundamentals for a color deficient observer of a given type.

Allowed types:

:protanope

:deuteranope

:tritanope

:protanomalous

:deuteranomalous

:tritanomalous

SpectralVis.colv — Method

colv(type,T,colmatch,s)

type: color space

T: whitepoint correlated color temperature

colmatch: color matching function

dl: delta lambda

output: a,b,c (locus vectors in given color space)

    x,y,Y for Yxy

    X,Z,Y for XYZ

    a,b,L for Luv, Lab, DIN99, DIN99d, DIN99o

SpectralVis.conefund — Method

conefund(g::CMatch)

Returns the CIE cone fundamentals associated with a CMF.

Possible types for g:

CIE31, CIE1931 2° observer

CIE64, CIE1964 10° observer

CIE12_2, CIE2012 2° observer (current CIE 2° standard observer)

CIE12_10, CIE2012 10° observer (current CIE 10° standard observer)

Example

julia> conefund(cmfunc(:cie31))
LMS31(Real[360, 361, 362, 363, 364, 365, 366, 367, 368, 369  …  821, 822, 823, 824, 825, 826, 827, 828, 829, 830], Real[6.209077859e-6, 6.950775712086998e-6, 7.782164998607999e-6, [...], 5.5398906285000006e-8, 5.1646151529000004e-8, 4.814798721600001e-8, 4.488794319300001e-8, 4.1849549307000004e-8])

SpectralVis.conefund — Method

conefund(g::Symbol)

Returns the CIE cone fundamentals according to the requested lms function.

Possible symbols for g:

:lms31, CIE1931 2° observer

:lms64, CIE1964 10° observer

:lms06_2, CIE2012 2° observer (current CIE 2° standard observer)

:lms06_10, CIE2012 10° observer (current CIE 10° standard observer)

Example

julia> conefund(:cie64)
LMS64(Real[360, 361, 362, 363, 364, 365, 366, 367, 368, 369  …  821, 822, 823, 824, 825, 826, 827, 828, 829, 830], Real[1.5474770176000004e-8, 2.3435133003999998e-8, 3.5282213430000004e-8, [...], 1.4321467274333274e-9, 1.384067146347377e-9, 1.3377105658218067e-9, 1.2930114805124651e-9, 1.2499070138616995e-9, 1.208336807203753e-9])

SpectralVis.coneresponse — Function

coneresponse(response::Symbol)

Return cone response matrix according to chromatic adaptation transform response.

Available transforms:

:xyzscaling Simplest (and worst) method.

:vonkries

:bradford Linearized Bradfod transform. Diagonal Matrix Transform (DMT)-type vonKries-Ives transform.

:sharp Sharp Adaptation Transform. First described by Finlayson et al.

:fairchild

:cat97s CIECAM97s adaptation transform

:cat02 Default transform. Developed for the CIECAM02 color appearance model.

:hpe Hunt-Pointer-Estevez transform

:bestof20 by Sabine Süsstrunk: Computing Chromatic Adaptation, PhD thesis, 2005, page 104. Sharpened linear cone response matrix optimized for white-point independence, gamut coverage, and hue constancy. Delivers better results than :cat02 in most cases. Added for experimental purposes.

Examples

julia> coneresponse(:xyzscaling)
3×3 Matrix{Float64}:
 1.0  0.0  0.0
 0.0  1.0  0.0
 0.0  0.0  1.0

julia> coneresponse(:bradford)
3×3 Matrix{Float64}:
  0.8951   0.2664  -0.1614
 -0.7502   1.7135   0.0367
  0.0389  -0.0685   1.0296

julia> coneresponse(:bestof20)
3×3 Matrix{Float64}:
  1.6351  -0.4071  -0.228
 -0.8044   1.7798   0.0246
  0.0      0.0152   0.9848

SpectralVis.convmat — Method

convmat(g::Symbol)

Returns the conversion matrix to convert CMFs to their associated cone fundamentals (LMS).

Result: 3 × 3 matrix

Example

convmat(:cie12_10)

3×3 Matrix{Float64}: 1.93986 -1.34664 0.430449 0.692839 0.349676 0.0 0.0 0.0 2.14688

SpectralVis.deactivate_cone — Method

deactivate_cone(cmf::ConeFund, conetype::Symbol

returns cone fundamental either l, m, or s cone function deactivated.

cone types: :l, :m, :s

Result: CMatch

SpectralVis.deactivate_cone — Method

deactivate_cone(conetype::Symbol

returns cone fundamental either l, m, or s cone function deactivated.

cone types: :l, :m, :s

Result: CMatch

SpectralVis.draw_ellipses — Method

draw_ellipses(dataset)

Draws all ellipses of a chosen dataset.

SpectralVis.ellipse — Method

ellipse(u, v, a, b, theta)

Create coordinate vectors of an ellipse with center coordinates u, v, major axis a, minor axis b, and tilt angle θ.

Result: x::Vector, y::vector

SpectralVis.ellipse_a — Method

ellipse_a(dataset, index::Int)

Gives the major axis parameter a for a dataset at a given index.

SpectralVis.ellipse_b — Method

ellipse_b(dataset, index::Int)

Gives the minor axis parameter b for a dataset at a given index.

SpectralVis.ellipse_center — Method

ellipse_center(dataset, index::Int)

Gives the pair of coordinates x0, y0 of the ellipse center for a dataset at a given index.

SpectralVis.ellipse_g11 — Method

ellipse_g11(dataset, index::Int)

Gives the ellipse parameter g11 at a given index.

SpectralVis.ellipse_g12 — Method

ellipse_g12(dataset, index::Int)

Gives the ellipse parameter g12 at a given index.

SpectralVis.ellipse_g22 — Method

ellipse_g22(dataset, index::Int)

Gives the ellipse parameter g22 at a given index.

SpectralVis.ellipse_g23 — Method

ellipse_g23(dataset, index::Int)

Gives the ellipse parameter g23 at a given index.

SpectralVis.ellipse_luminance — Method

ellipse_luminance(dataset,index::Int)

Gives the luminance in cd/m² of a given test color

SpectralVis.ellipse_parameters — Method

ellipse_parameters(dataset, index::Int)

Gives the entire set of ellipse parameters for a dataset at a given index.

SpectralVis.ellipse_θ — Method

ellipse_b(dataset, index::Int)

Gives the axis tilt parameter θ for a dataset at a given index.

SpectralVis.ellispe_g13 — Method

ellipse_g13(dataset, index::Int)

Gives the ellipse parameter g13 at a given index.

SpectralVis.ellispe_g33 — Method

ellipse_g33(dataset, index::Int)

Gives the ellipse parameter g33 at a given index.

SpectralVis.fundamental_metamer — Function

fundamental_metamer(illum::LSpec, rad::Spectrum, match::ConeFund = conefund(SPECENV.cmf),type::Symbol = :standard)

Creates fundamental metamer of a spectrum according to Cohen’s matrix R method.

Examples:

SpectralVis.gamut_vis — Function

gamut_vis(type,T1::Real,dT::Real,T2::Real,colmatch,name::AbstractString,altitude::Real,azimuth::Real,slice=collect(5:5:95))

Creates gamut given by Temperature range

SpectralVis.gamut_vis — Function

gamut_vis(type, title, lum, whitepoint, colmatch,name::AbstractString,altitude::Real,slice=collect(5:5:95))

creates 3D flyaround of gamut given by illuminant

Example: gamut_vis(Colors.DIN99o,"DIN99o at 6500K", D_series_luminance(6500),D_series_whitepoint(6500,CMF2012_10),CMF2012_10,"DIN99o_6500K_",35,collect(5:5:95))

SpectralVis.gas_discharge_illuminant — Method

gas_discharge_illuminant(n::Int)

Returns gas discharge type illuminant of index n for n from 1 to 7.

Result: ::LSpec

SpectralVis.gas_discharge_illuminant — Method

gas_discharge_illuminant(type::Symbol)

SPDs of common gas discharge lamps.

available types:

:lps: Low pressure sodium

:hps: High pressure sodium

:mb: High pressure mercury

:mbf: High pressure mercury with red-emitting phosphor coating, better CRI than MB

:mbtf: High pressure mercury with red-emitting phopsphor coating and tungsten filament ballast, better CRI

:hmi: High pressure mercury with metal halides. Medium arc, iodides, typical stadium lighting, was used in television to supplement daylight

:xenon: Xenon, used in film projectors in cinemas, floodlighting, in lighthouses, for accelerated fading tests, flash photography, lighting fluorescent materials

Example

julia> gas_discharge_illuminant(:lps)
LSpec(Real[390.0, 391.0, 392.0, 393.0, 394.0, 395.0, 396.0, 397.0, 398.0, 399.0  …  821.0, 822.0, 823.0, 824.0, 825.0, 826.0, 827.0, 828.0, 829.0, 830.0], Real[0.0, 0.02, 0.04, 0.06, 0.08, 0.1, 0.08, 0.060000000000000005, 0.04000000000000001, 0.020000000000000004  …  0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0])

julia> gas_discharge_illuminant(:xenon)
LSpec(Real[390.0, 391.0, 392.0, 393.0, 394.0, 395.0, 396.0, 397.0, 398.0, 399.0  …  821.0, 822.0, 823.0, 824.0, 825.0, 826.0, 827.0, 828.0, 829.0, 830.0], Real[96.33, 97.176, 98.022, 98.868, 99.714, 100.56, 101.01, 101.46000000000001, 101.91, 102.36  …  0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0])

SpectralVis.hardlimit_spec — Function

`hardlimit_spec(s::Spectrum,limit::Real = 1.0)`

Hardlimits an SPD to the maximum value limit.

Default limit is 1.0

Examples

julia> hardlimit_spec(D50_illuminant())
LSpec(Real[390.0, 391.0, 392.0, 393.0, 394.0, 395.0, 396.0, 397.0, 398.0, 399.0  …  821.0, 822.0, 823.0, 824.0, 825.0, 826.0, 827.0, 828.0, 829.0, 830.0], Real[1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0  …  1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0])

julia> hardlimit_spec(D_series_illuminant(6500),59.0)
LSpec(Real[390.0, 391.0, 392.0, 393.0, 394.0, 395.0, 396.0, 397.0, 398.0, 399.0  …  821.0, 822.0, 823.0, 824.0, 825.0, 826.0, 827.0, 828.0, 829.0, 830.0], Real[54.60212796545921, 57.41151658372402, 59.0, 59.0, 59.0, 59.0, 59.0, 59.0, 59.0, 59.0  …  57.744482738230694, 58.03176686777033, 58.319050997309965, 58.6063351268496, 58.893619256389236, 59.0, 59.0, 59.0, 59.0, 59.0])

SpectralVis.illuminant_A — Function

illuminant_A(env::SpecEnv = SPECENV)

Produces the spectrum of illuminant A, which represents tungsten-filament lighting, depending on the limits defined by SPECENV. The relative spectral power distribution is equal to a Planckian spectrum at T ≊ 2856 K, which is also its CCT.

CCT according to the International Practical Temperature Scale from 1968.

This is the illuminant that should always be used when incandescent lighting is involved.

Example

julia> illuminant_A()
LSpec(Real[390.0, 391.0, 392.0, 393.0, 394.0, 395.0, 396.0, 397.0, 398.0, 399.0  …  821.0, 822.0, 823.0, 824.0, 825.0, 826.0, 827.0, 828.0, 829.0, 830.0], Real[1.0161486323473047e11, 1.0369227145593657e11, 1.05797774054285e11, 1.0793148447599527e11, 1.1009351284197664e11, 1.1228396592637178e11, 1.1450294713618913e11, 1.1675055649202544e11, 1.1902689060987242e11, 1.2133204268399858e11  …  2.17270715240179e12, 2.1757420157602646e12, 2.178757889232851e12, 2.1817547874185107e12, 2.1847327254181292e12, 2.1876917188295703e12, 2.190631783742765e12, 2.1935529367348203e12, 2.196455194865146e12, 2.199338575670583e12])

SpectralVis.illuminant_B — Function

illuminant_B(env::SpecEnv = SPECENV)

Produces the spectrum of CIE Standard Illuminant B, with a CCT of ≈4874 K. Originally intended to represent sunlight. Now obsolete, but according to Hunt the European brewing industry still uses an approximation of this illuminant.

Example

julia> illuminant_B()
LSpec(Real[390.0, 391.0, 392.0, 393.0, 394.0, 395.0, 396.0, 397.0, 398.0, 399.0  …  821.0, 822.0, 823.0, 824.0, 825.0, 826.0, 827.0, 828.0, 829.0, 830.0], Real[0.313, 0.3223587343061588, 0.3319703445293299, 0.3417825875994216, 0.3517432204463422, 0.3618, 0.3719111024629742, 0.38207638112852754, 0.3923061085625939, 0.4026105573311068  …  7.557164594645486e-6, 9.446455743306857e-6, 7.557164594645484e-6, 3.778582297322749e-6, 1.734723475976807e-20, -2.267149378393646e-6, -3.0228658378581947e-6, -2.6450076081259192e-6, -1.5114329189290965e-6, -2.168404344971009e-21])

SpectralVis.illuminant_C — Function

illuminant_C(env::SpecEnv)

Produces the spectrum of CIE Standard Illuminant C, with a CCT of ≈ 6774 K. Originally intended to represent average daylight.

Example

julia> illuminant_C()
LSpec(Real[390.0, 391.0, 392.0, 393.0, 394.0, 395.0, 396.0, 397.0, 398.0, 399.0  …  821.0, 822.0, 823.0, 824.0, 825.0, 826.0, 827.0, 828.0, 829.0, 830.0], Real[0.474, 0.489271743700384, 0.5046749637120697, 0.5202123118735634, 0.5358864400233714, 0.5517, 0.5676560269088902, 0.583759088923221, 0.6000141374831071, 0.616426124028662  …  0.7121636730114121, 0.7159995912642652, 0.7194836730114121, 0.722791836505706, 0.7261, 0.7295488980965763, 0.7331385307954352, 0.7368337144460056, 0.7405992653977175, 0.7444])

How the CIE Standard Illuminant C is produced:

Standard Source A is combined with a filter made of two 1 cm thick layers, containing two solutions, C₁ and C₂. The solutions are contained in a double cell that is made of a colorless optical glass.

Solution C₁:

3.412 g copper sulphate ($CuSO_4·5H_2O$)

3.412 g mannite [$C_6H_8(OH)_6$]

30.0 ml pyridine ($C_5H_5N$)

1000.0 ml distilled water

Solution C₂:

30.58 g cobalt ammonium sulphate [$CoSO_4·(NH_4)_2SO_4·6H_2O$]

22.52 g copper sulphate ($CuSO_4·5H_2O$)

10.0 ml sulphuric acid (density 1.835 g/ml)

1000.0 ml distilled water

SpectralVis.inglingtsou_opponent — Method

inglingtsou_opponent(conefund::ConeFund)

Converts cone fundamentals to color opponent channels using the Ingling-Tsou transformation matrix.

Signal flow of the two-stage model of human color vision:

    L              M                    S   photoreceptor
    ┃              ┃                    ┃      stage
    ┃              ┃     ┏━━━━━━━━━━━━━━┫      (CMFs)
    ┃  ┏━━━━━━━━━━━╋━━━━━┃━━━━━━━━━┓    ┃
    ┃  ┃           ┃     ┃         ┃    ┃
 ┏━━┻━━┃━━━━━┳━━━━━┃━━━━━┃━━━━┓    ┃    ┃
 ┃     ┃     ┃     ┃     ┃    ┃    ┃    ┃
 ┃     ┃     ┃     ┃     ┃    ┃    ┃    ┃
 ┃     ┃     ┃     ┃     ┃    ┃    ┃    ┃       
0.6   0.4  0.24  0.105  0.7  1.2  1.6  0.4   factors
 ║     ║     ║     ║     ║    ║    ║    ║
 ╚══╦══╝     ╚═════╬═════╝    ╚════╬════╝
    ║              ║               ║
    ║              ║               ║
 L  +  M     L  +  M  -  S    L  - M +  S     opponent
    ║              ║               ║           stage
    Vλ            y-b             r-g

lightness     yellow-blue       red-green

Example

julia> inglingtsou_opponent(conefund(SPECENV.cmf))
VYBRG(Real[390.0, 391.0, 392.0, 393.0, 394.0, 395.0, 396.0, 397.0, 398.0, 399.0  …  821.0, 822.0, 823.0, 824.0, 825.0, 826.0, 827.0, 828.0, 829.0, 830.0], Real[0.0003878623982, 0.0004737047352, 0.000577319691, 0.0007018085474, 0.0008506078488000001, 0.0010274557607999999, 0.001236329446, 0.0014813467259999998, 0.001766625772, 0.0020960976039999996  …  9.411752703999999e-7, 8.876810727999999e-7, 8.372313131999999e-7, 7.896833636e-7, 7.448974792000001e-7, 7.027372568e-7, 6.630699992e-7, 6.257670254799999e-7, 5.9070389244e-7, 5.5776058904e-7], Real[-0.004164411112444999, -0.00504460427438, -0.006110193693075, -0.007397180969524999, -0.00894723637031, -0.010808111330739998, -0.013033902761799998, -0.01568509182405, -0.018828265622099997, -0.02253541766325  …  3.6881237642999995e-7, 3.4782318591e-7, 3.2803082274e-7, 3.0937903046999996e-7, 2.9181270796499996e-7, 2.7527808831e-7, 2.5972288066499997e-7, 2.450963909835e-7, 2.31349607448e-7, 2.18435275848e-7], Real[0.0023730382284000004, 0.0028717459496, 0.003474333342, 0.00420068739, 0.0050737833632, 0.0061199115808, 0.007368828876, 0.008853794675999999, 0.010611444512, 0.012681445520000003  …  1.5481949744e-6, 1.4590315327999997e-6, 1.3750428792e-6, 1.2959784376e-6, 1.2215934872e-6, 1.1516498848e-6, 1.0859166232e-6, 1.02417030368e-6, 9.661954418399999e-7, 9.117847518400001e-7])

SpectralVis.led — Method

LED helper function

SpectralVis.led_spec — Method

led_spec(λmin, λmax, Δλ, λ0, Δλ1_2)

LED spectrum simulation

λ0: peak wavelength

Δλ1_2: half spectral width, width at 50% peak intensity

Result: LSpec

Source: Yoshi Ohno (2004), Simulation Analysis of White LED Spectra and Color Rendering, Proceedings for CIE Expert Sumposium on LED Light Sources, Tokyo, JA (Accessed May 27, 2021)

SpectralVis.luminance_spec — Method

luminance_spec(λ,r)`

Luminance spectrum of type RSpec

SpectralVis.macbeth — Method

macbeth(colorchart::Symbol,num::Int)

returns RSpec of one of the 24 MacBeth color patches of a given number.

available color charts:

:Macbeth_chart1, λ=380:10:730 nm, average of multiple measured charts from BabelColor

:Macbeth_chart2, λ=380:4:780 nm

:Macbeth_chart3, λ=380:5:780 nm

available color patch numbers: 1 to 24

SpectralVis.macbeth — Method

macbeth(colorchart::Symbol,color::Symbol)

returns RSpec of one of the 24 MacBeth color patches of a given name.

available color charts:

:Macbeth_chart1, λ=380:10:730 nm, average of multiple measured charts from BabelColor

:Macbeth_chart2, λ=380:4:780 nm

:Macbeth_chart3, λ=380:5:780 nm

available color names:

:darkskin, :lightskin, :bluesky, :foliage, :blueflower, :bluishgreen,

:orange, :purplishblue, :moderatered, :purple, :yellowgreen, :orangeyellow,

:blue, :green, :red, :yellow, :magenta, :cyan,

:white95, :neutral8, :neutral65, :neutral5, :neutral35, :black2

SpectralVis.matrixA — Function

matrixA(match::ConeFund = conefund(SPECENV.cmf), type::Symbol = :standard)

Creates matrix A from a cone fundamental for Cohen’s matrix R calculations.

Available types:

:protanopia

:deuteranopia

:tritanopia

:standard

Examples:

julia> matrixA(conefund(SPECENV.cmf))
441×3 Matrix{Float64}:
 0.0108549   0.00218811  0.028312
 0.0131472   0.00265254  0.0343045
 0.0159195   0.00321355  0.0415584
 0.0192642   0.00388931  0.0503177
[...]
 2.6301e-6   1.56603e-6  0.0
 2.48078e-6  1.47742e-6  0.0
 2.34056e-6  1.39418e-6  0.0
 2.20894e-6  1.31601e-6  0.0

SpectralVis.matrixR — Function

matrixR(match::ConeFund = conefund(SPECENV.cmf), type::Symbol = :standard)

Creates Cohen’s matrix R from a given CMF. For available types, see matrixA.

Examples:

julia> matrixR(conefund(SPECENV.cmf), :deuteranopia)
441×441 Matrix{Float64}:
 9.97123e-7   1.20808e-6   1.4634e-6    1.77166e-6   …  4.61275e-11  4.35087e-11  4.10494e-11  3.8741e-11
 1.20808e-6   1.46367e-6   1.77301e-6   2.14648e-6      5.57733e-11  5.26069e-11  4.96333e-11  4.68422e-11
 1.4634e-6    1.77301e-6   2.14773e-6   2.60013e-6      6.73948e-11  6.35686e-11  5.99755e-11  5.66027e-11
[...]
 4.35087e-11  5.26069e-11  6.35686e-11  7.6741e-11      5.56939e-14  5.2532e-14   4.95627e-14  4.67755e-14
 4.10494e-11  4.96333e-11  5.99755e-11  7.24034e-11     5.25459e-14  4.95627e-14  4.67613e-14  4.41316e-14
 3.8741e-11   4.68422e-11  5.66027e-11  6.83317e-11  …  4.95909e-14  4.67755e-14  4.41316e-14  4.16498e-14

SpectralVis.mbcolor — Method

mbcolor(spec::RSpec,illum::LSpec)

returns XYZ color according to illuminant

SpectralVis.mbspec — Method

mbspec(colorchart::Symbol, num::Int)

returns RSpec adapted to th 390:830 nm range of the standard CMF

SpectralVis.mbspec — Method

mbspec(colorchart::Symbol,name::Symbol)

returns RSpec adapted to SPECENV limits for λ.

SpectralVis.mer_caltarget — Method

mer_caltarget(color::Int)

Returns the spectrum of a given MER Caltarget color patch.

Allowed color patch indices: 1 ≤ color ≤ 7.

Result: RSpec

SpectralVis.mer_caltarget — Method

mer_caltarget(color::Symbol)

Returns the spectrum of a given MER Caltarget color patch.

Allowed color patch names:

:white, :grey, :black, :yellow, :red, :green, :blue

Result: RSpec

SpectralVis.metameric_black — Function

metameric_black(illum::LSpec, rad::Spectrum, match::ConeFund = conefund(SPECENV.cmf), type::Symbol = :standard)

Creates metameric black according to Cohen’s matrix R method.

Examples

SpectralVis.mineral — Function

mineral(mineral::Symbol, spectable = mineral_specs(mineral))

Returns reflectance spectrum of chosen mineral.

Allowed names for mineral: :augite, :goethite, :hematite, :pigeonite

SpectralVis.mineral_specs — Method

mineral_specs(mineral::Symbol)

Reads and returns the spectral data of a chosen mineral.

Available mineral types:

:augite

:goethite

:hematite

:pigeonite

SpectralVis.munsell_spec — Function

munsell_spec(idx::Int, speclist)

Returns reflectance spectrum of the Munsell dataset at index idx. Possible indices go from 1 to 1269

Example

julia> munsell_spec(437, speclist)
RSpec(Real[380, 381, 382, 383, 384, 385, 386, 387, 388, 389  …  791, 792, 793, 794, 795, 796, 797, 798, 799, 800], Real[0.1184, 0.1208, 0.124, 0.1301, 0.1349, 0.1431, 0.1498, 0.1579, 0.1631, 0.1731  …  0.3355, 0.3364, 0.3345, 0.3337, 0.3327, 0.3348, 0.3392, 0.3377, 0.335, 0.3351])

SpectralVis.munsell_specs — Method

munsell_specs()

Returns an array containing all 1269 munsell spectra, measured from 380 to 800 nm in 1 nm steps.

SpectralVis.normalize_D_series — Method

normalize_D_series(s::Spectrum)

normalize a D_series spectrum to 1.0 at 560 nm.

SpectralVis.normalize_blackbody — Method

`normalize_blackbody(s::Spectrum)`

Normalize a blackbody spectrum analogous to D series to 1.0 at 560 nm.

SpectralVis.normalize_spec — Method

normalize_spec(s::Spectrum,value::Real)

returns a spectrum with reflectance distribution normalized to value relative to max.

SpectralVis.normalize_spec — Method

normalize_spec(s::Spectrum)

returns a spectrum with reflectance values normalized to 1.0

SpectralVis.normalized_D_series — Method

normalized_D_series(env::SpecEnv = SPECENV, T::Real)

produces CIE Daylight spectrum according to desired CCT, normalized to l=100 at λ=560 nm.

Examples

julia> normalized_D_series(SPECENV, 6500)
LSpec(Real[390.0, 391.0, 392.0, 393.0, 394.0, 395.0, 396.0, 397.0, 398.0, 399.0  …  821.0, 822.0, 823.0, 824.0, 825.0, 826.0, 827.0, 828.0, 829.0, 830.0], Real[0.5460212796545921, 0.5650827295052294, 0.590288902296674, 0.620011293254124, 0.6526213976027774, 0.6864907105678326, 0.7200645866061678, 0.7520838171013811, 0.7813630526687508, 0.8067169439235554  …  0.578028743266789, 0.5810475635333059, 0.5837744258575818, 0.5863553092107371, 0.5889361925638924, 0.5916338590939442, 0.5944483088008923, 0.5973503458905127, 0.6003107745685814, 0.6033003990408742])

normalized_D_series(9300)
LSpec(Real[390.0, 391.0, 392.0, 393.0, 394.0, 395.0, 396.0, 397.0, 398.0, 399.0  …  821.0, 822.0, 823.0, 824.0, 825.0, 826.0, 827.0, 828.0, 829.0, 830.0], Real[0.9296442162075731, 0.9527892027932645, 0.9854018656135435, 1.0248615127313507, 1.0685474522096268, 1.1138389921113123, 1.158318458802078, 1.2003802518585145, 1.2386217891599418, 1.2716404885856807  …  0.4869584146523383, 0.4894740470571454, 0.4917569270423351, 0.4939234308177161, 0.4960899345930971, 0.49834953933632503, 0.5007022450474, 0.5031247764843602, 0.5055938584052438, 0.5080862155680892])

SpectralVis.reflectance_spec — Method

reflectance_spec(λ,r)

Reflectance spectrum of type RSpec

SpectralVis.reinterpolate — Function

reinterpolate(spectrum,λmin,Δλ,λmax,style=:natural)

SpectralVis.set_colormatch — Method

set_colormatch(env::SpecEnvironment, colormatch::CMatch)

Sets the color matching function (CMF) used in a defined spectral environment.

Available CMFs:

:cie31, :cie31_j, :cie31_jv, :cie64, :cie12_2, :cie12_10, :lms06_2, :lms06_10

SpectralVis.set_colormatch — Method

set_colormatch(colormatch::CMatch)

Sets the color matching function (CMF) used in the spectral environment.

Available CMFs:

:cie31, :cie31_j, :cie31_jv, :cie64, :cie12_2, :cie12_10, :lms06_2, :lms06_10

SpectralVis.set_extrap — Method

set_extrap(env::SpecEnvironment,extrapolation::Symbol)

Sets the extrapolation mode of the environment env to one of the available extrapolation settings:

:none or :zero: out of range values are zero.

:constant: out of range values are set to λmin and λmax, respectively.

:linear: out of range values are linearly extrapolated through the first/last two data points.

:parabolic: out fo range values are extrapolated with a parabola through the first/last three datapoints. Recommended extrapolation method according to D.L. MacAdams in "Color Measurement", chapter 5.4 "Truncations".

SpectralVis.set_extrap — Method

set_extrap(extrapolation::Symbol)

Sets the extrapolation mode of the environment env to one of the available extrapolation settings:

:none or :zero: out of range values are zero.

:constant: out of range values are set to λmin and λmax, respectively.

:linear: out of range values are linearly extrapolated through the first/last two data points.

:parabolic: out fo range values are extrapolated with a parabola through the first/last three datapoints. Recommended extrapolation method according to D.L. MacAdams in "Color Measurement", chapter 5.4 "Truncations".

SpectralVis.set_limits — Method

set_limits(λmin::Real,Δλ::Real,λmax::Real)

sets the wavelength limits and resolution of the spectral environment

SpectralVis.set_limits — Method

set_limits(env::SpecEnvironment, λmin::Real,Δλ::Real,λmax::Real)

sets the wavelength limits and resolution of the spectral environment

SpectralVis.set_specenv — Function

set_specenv(λmin=390.0,Δλ=1.0,λmax=830.0,extrapolate= :zero)

Initializes the spectral environment to standard values:

λmin = 380.0 nm λmax = 830.0 nm Δλ = 1.0 nm cmf = cmfunc(:cie12_10) ex = :linear

SpectralVis.shift_cone — Method

shift_colormatch(cmf::CMatch, conetype::Symbol, Δλ)

returns cone fundamental associated with the given CMF, with shifted cone response function l, m, or s

cone types: :l, :m, :s

Result: CMatch

SpectralVis.shift_cone — Method

shift_colormatch(cmf::ConeFund,conetype::Symbol,Δλ)

returns cone fundamental with shifted cone response function l, m, or s.

cone types: :l, :m, :s

Result: CMatch

SpectralVis.shift_spec — Method

shift_spec(spec::Spectrum,Δλ)

Returns spectrum of same type, shifted by Δλ

SpectralVis.sinusoidal_spd — Function

sinusoidal_spd(freq::Real, modfac::Real = 1.0, ampl::Real = 1.0, phase::Real, r = (SPECENV.λmin:SPECENV.Δλ:SPECENV.λmax))

Generates a sinusoidal spectral power distribution. Useful for computational gamut and spectral calculations.

SpectralVis.skin_simulation — Method

skin_simulation()

Skin optics Simulation

SpectralVis.sl — Method

sl(type,lum::LSpec,whitepoint,colmatch,s,d,e)

SpectralVis.spec_extrapolate — Function

spline_extrap(Spl, env=SPECENV)

extrapolation of a spline according to spectral environment.

SpectralVis.transmittance_spec — Method

transmittance_spec(λ,t)

Tansmittance spectrum at unit thickness (x = 1.0).

SpectralVis.transmittance_spec — Method

transmittance_spec(λ,t,d,d0)

Transmittance spectrum with thickness term d. Unit thickness is d0, according to Bouguer’s Law/Lambert’s Law.

See: "Color Measurement" by David L. MacAdam, chapter 3.2: "Bougouer’s Law"

SpectralVis.transmittance_spec — Method

transmittance_spec(λ, t, x)

Transmittance spectrum with thickness x given relative to unit thickness, according to Bougouer’s Law/Lambert’s Law.

See: "Color Measurement" by David L. MacAdam, chapter 3.2: "Bouguer’s Law"

SpectralVis.tristimulus — Function

tristimulus(illum::LSpec, rad::Spectrum, match::ConeFund = conefund(SPECENV.cmf), type::Symbol = :standard)

Creates tristimulus values resulting from radiometric function/stimulus reaching the eye, according to Cohen’s matrix R method.

Examples:

julia> tristimulus(D65_illuminant(), adapt_spec(agfait(100), SPECENV, :zero, :cubic, :natural)) 3-element Vector{Any}: 2827.9824977740263 4605.154579468038 6719.982399962246`

SpectralVis.vybrg — Function

vybrg(conefund::ConeFund = conefund(SPECENV.cmf))

Returns the Ingling-Tsou opponent color functions based on the cone fundamental functios.

Example

julia> c = conefund(:cie12_2); vybrg(c)
VYBRG(Real[390.0, 391.0, 392.0, 393.0, 394.0, 395.0, 396.0, 397.0, 398.0, 399.0  …  821.0, 822.0, 823.0, 824.0, 825.0, 826.0, 827.0, 828.0, 829.0, 830.0], Real[0.0003963417152, 0.0004807961832, 0.0005820062634, [...], 1.4450193672e-6, 1.3620748624e-6, 1.2840832415999998e-6,] 1.2107868407999998e-6, 1.1419364999999999e-6, 1.0772919015999998e-6, 1.0166218466399999e-6])

SpectralVis.wratten — Method

wratten(f::Int)

Returns `TSpec` of one of the Wratten filters of a given index `f` in the range `0 < f < 128`.

Examples

julia> wratten(77)
TSpec(Real[400.0, 410.0, 420.0, 430.0, 440.0, 450.0, 460.0, 470.0, 480.0, 490.0  …  610.0,
 620.0, 630.0, 640.0, 650.0, 660.0, 670.0, 680.0, 690.0, 700.0], Real[0.6409999999999999,
  0.665, 0.684, 0.695, 0.7, 0.6940000000000001, 0.675, 0.654, 0.629, 0.598  …  0.4, 0.376,
   0.355, 0.342, 0.336, 0.34, 0.341, 0.345, 0.348, 0.36], 1.0)

SpectralVis.wratten — Method

wratten(f::Symbol)

Returns TSpec of one of the Wratten filters of a given name f.

Examples

julia> wratten(:n44)
TSpec(Real[400.0, 410.0, 420.0, 430.0, 440.0, 450.0, 460.0, 470.0, 480.0, 490.0  …  610.0, 
620.0, 630.0, 640.0, 650.0, 660.0, 670.0, 680.0, 690.0, 700.0], Real[0.0044, 0.0036, 0.0063,
 0.0363, 0.131, 0.254, 0.365, 0.465, 0.536, 0.568…  0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0,
  0.0018, 0.016], 1.0)

SpectralVis.wratten_filter — Function

wratten_filter(f::Int, x::Real)

Returns TSpec of one of the Wratten filters of a given name adapted to SPECENV settings. x is relative to unit thickness.

SpectralVis.wratten_filter — Method

wratten_filter(f::Int)

returns TSpec of one of the Wratten filters at unit thickness.

SpectralVis.wratten_filter — Method

wratten_filter(f::Symbol)

returns TSpec of one of the Wratten filters of a given name adapted to SPECENV settings.

SpectralVis.wratten_filter — Method

wratten_filter(f::Symbol)

returns TSpec of one of the Wratten filters of a given name at unit thickness x = 1.0

SpectralVis.CIE12_10 — Type

CIE12_10

Type of CIE 2012 10° observer color matching function. Current CIE 10° standard observer.

SpectralVis.CIE12_2 — Type

CIE12_2

Type of CIE 2012 2° observer color matching function. Current CIE 2° standard observer.

SpectralVis.CIE31 — Type

CIE31

Type of CIE 1931 2° observer color matching function.

SpectralVis.CIE31_J — Type

CIE31_J

Type of CIE 1931 2° observer color matching function with Judd corrections.

SpectralVis.CIE31_JV — Type

CIE31_JV

Type of CIE 1931 2° observer color matching function with corrections by Judd and Vos.

SpectralVis.CIE64 — Type

CIE31_J

Type of CIE 1964 10° observer color matching function.

SpectralVis.LMS06_10 — Type

LMS06_10

Type of CIE 2006 10° observer cone fundamentals.

SpectralVis.LMS06_2 — Type

LMS06_2

Type of CIE 2006 2° observer cone fundamentals.

SpectralVis.LMS31 — Type

LMS31

Type of CIE 1931 2° observer cone fundamentals.

SpectralVis.LMS64 — Type

LMS64

Type of CIE 1964 10° observer cone fundamentals.

SpectralVis.LSpec — Type

LSpec, Luminance Spectrum

Fields:

λ::Vector{Real}, vector containing wavelengths

l::Vector{Real}, vector containing luminance values

SpectralVis.RSpec — Type

RSpec, Reflectance Spectrum

Fields:

λ::Vector{Real}, vector containing wavelengths

`r::Vector{Real}, vector containing reflectanceS values

SpectralVis.TSpec — Type

TSpec, Transmittance spectrum

Fields:

λ::Vector{Real}, vector containing wavelengths

t::Vector{Real}, vector containing transmittance values

x::Real, unit thickness value

SpectralVis.VYBRG — Type

VYBRG

Type of Ingling-Tsou color opponent space functions with Vλ (lightness), yb (yellow-blue), rg (red-green) color channels.