Spectral Data

SpectralVis comes with a large variety of spectral datasets.

Agfa IT8 target

agfait_specs()

SpectralVis.agfait_specs — Function

agfait_specs()

agfait(idx::Int)

SpectralVis.agfait — Function

agfait(idx::Int)

Returns the reflectance spectrum of the selected Agfait872 color patch. Spectral data are from 400 to 700 nm in 10 nm steps.

1st group: 1-264

2nd group: Neutral scale samples (265-286)

3rd group:

black sample (287)

white sample (288)

Minolta white calibration sample (289)

Examples

julia> agfait(123)
RSpec(Real[400.0, 410.0, 420.0, 430.0, 440.0, 450.0, 460.0, 470.0, 480.0, 490.0  …  610.0, 620.0, 630.0, 640.0, 650.0, 660.0, 670.0, 680.0, 690.0, 700.0], Real[0.1282, 0.3004, 0.4806, 0.5796, 0.6429, 0.6818000000000001, 0.7078, 0.7159, 0.7071999999999999, 0.6896  …  0.259, 0.2318, 0.21309999999999998, 0.2, 0.1932, 0.19210000000000002, 0.19699999999999998, 0.2095, 0.2311, 0.2594])

julia> agfait(0)
ERROR: DomainError with 0:
Color index out of range. Expecting 0 ≤ index ≤ 289.

Measurer: ela@ee.oulu.fi (University of Oulu, Electrical Eng.Dept)

Munsell matte spectra

munsell_specs()

SpectralVis.munsell_specs — Function

munsell_specs()

Returns an array containing all 1269 munsell spectra, measured from 380 to 800 nm in 1 nm steps.

munsell_spec(idx::Int, speclist)

SpectralVis.munsell_spec — Function

munsell_spec(idx::Int, speclist)

Returns reflectance spectrum of the Munsell dataset at index idx. Possible indices go from 1 to 1269

Example

julia> munsell_spec(437, speclist)
RSpec(Real[380, 381, 382, 383, 384, 385, 386, 387, 388, 389  …  791, 792, 793, 794, 795, 796, 797, 798, 799, 800], Real[0.1184, 0.1208, 0.124, 0.1301, 0.1349, 0.1431, 0.1498, 0.1579, 0.1631, 0.1731  …  0.3355, 0.3364, 0.3345, 0.3337, 0.3327, 0.3348, 0.3392, 0.3377, 0.335, 0.3351])

MacBeth color chart

macbeth(colorchart::Symbol, color::Symbol)

SpectralVis.macbeth — Function

macbeth(colorchart::Symbol,color::Symbol)

returns RSpec of one of the 24 MacBeth color patches of a given name.

available color charts:

:Macbeth_chart1, λ=380:10:730 nm, average of multiple measured charts from BabelColor

:Macbeth_chart2, λ=380:4:780 nm

:Macbeth_chart3, λ=380:5:780 nm

available color names:

:darkskin, :lightskin, :bluesky, :foliage, :blueflower, :bluishgreen,

:orange, :purplishblue, :moderatered, :purple, :yellowgreen, :orangeyellow,

:blue, :green, :red, :yellow, :magenta, :cyan,

:white95, :neutral8, :neutral65, :neutral5, :neutral35, :black2

macbeth(colorchart::Symbol,num::Int)

returns RSpec of one of the 24 MacBeth color patches of a given number.

available color charts:

:Macbeth_chart1, λ=380:10:730 nm, average of multiple measured charts from BabelColor

:Macbeth_chart2, λ=380:4:780 nm

:Macbeth_chart3, λ=380:5:780 nm

available color patch numbers: 1 to 24

mbspec(colorchart::Symbol, name::Symbol)

SpectralVis.mbspec — Function

mbspec(colorchart::Symbol,name::Symbol)

returns RSpec adapted to SPECENV limits for λ.

mbspec(colorchart::Symbol, num::Int)

returns RSpec adapted to th 390:830 nm range of the standard CMF

mbcolor(spec::RSpec, illum::LSpec)

SpectralVis.mbcolor — Function

mbcolor(spec::RSpec,illum::LSpec)

returns XYZ color according to illuminant

Mars Exploration Rover Calibration Target (MER CalTarget)

mer_caltarget(color::Symbol)

SpectralVis.mer_caltarget — Function

mer_caltarget(color::Symbol)

Returns the spectrum of a given MER Caltarget color patch.

Allowed color patch names:

:white, :grey, :black, :yellow, :red, :green, :blue

Result: RSpec

mer_caltarget(color::Int)

Returns the spectrum of a given MER Caltarget color patch.

Allowed color patch indices: 1 ≤ color ≤ 7.

Result: RSpec

Mars Exploration Rover Panoramic Camera filters

pancam(filter::Symbol)

Missing docstring.

Missing docstring for pancam. Check Documenter's build log for details.

MER Opportunity measurement of the Martian Sky

mars_sky_opportunity()

Missing docstring.

Missing docstring for mars_sky_opportunity. Check Documenter's build log for details.

Kodak Wratten Filters

wratten(f::Symbol)

SpectralVis.wratten — Function

wratten(f::Int)

Returns `TSpec` of one of the Wratten filters of a given index `f` in the range `0 < f < 128`.

Examples

julia> wratten(77)
TSpec(Real[400.0, 410.0, 420.0, 430.0, 440.0, 450.0, 460.0, 470.0, 480.0, 490.0  …  610.0,
 620.0, 630.0, 640.0, 650.0, 660.0, 670.0, 680.0, 690.0, 700.0], Real[0.6409999999999999,
  0.665, 0.684, 0.695, 0.7, 0.6940000000000001, 0.675, 0.654, 0.629, 0.598  …  0.4, 0.376,
   0.355, 0.342, 0.336, 0.34, 0.341, 0.345, 0.348, 0.36], 1.0)

wratten(f::Symbol)

Returns TSpec of one of the Wratten filters of a given name f.

Examples

julia> wratten(:n44)
TSpec(Real[400.0, 410.0, 420.0, 430.0, 440.0, 450.0, 460.0, 470.0, 480.0, 490.0  …  610.0, 
620.0, 630.0, 640.0, 650.0, 660.0, 670.0, 680.0, 690.0, 700.0], Real[0.0044, 0.0036, 0.0063,
 0.0363, 0.131, 0.254, 0.365, 0.465, 0.536, 0.568…  0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0,
  0.0018, 0.016], 1.0)

wratten_filter(f::Int, x::Real)

SpectralVis.wratten_filter — Function

wratten_filter(f::Int, x::Real)

Returns TSpec of one of the Wratten filters of a given name adapted to SPECENV settings. x is relative to unit thickness.

wratten_filter(f::Int)

returns TSpec of one of the Wratten filters at unit thickness.

wratten_filter(f::Symbol)

returns TSpec of one of the Wratten filters of a given name adapted to SPECENV settings.

wratten_filter(f::Symbol)

returns TSpec of one of the Wratten filters of a given name at unit thickness x = 1.0

Vantablack®

vantablack()

Missing docstring.

Missing docstring for vantablack. Check Documenter's build log for details.

Mineral spectra

mineral_specs(mineral::Symbol)

SpectralVis.mineral_specs — Function

mineral_specs(mineral::Symbol)

Reads and returns the spectral data of a chosen mineral.

Available mineral types:

:augite

:goethite

:hematite

:pigeonite

mineral(mineral::Symbol, spectable = mineral_specs(mineral))

SpectralVis.mineral — Function

mineral(mineral::Symbol, spectable = mineral_specs(mineral))

Returns reflectance spectrum of chosen mineral.

Allowed names for mineral: :augite, :goethite, :hematite, :pigeonite