Spectral Environment

SpectralVis initializes a spectral environment which defines parameters that all spectral functions can access to guarantee compatibility of datasets and spectra.

SPECENV

The struct SPECENV stores all relevant parameters for spectral calculations. It contains the following fields:

λmin: wavelength of the lower boundary in nm.

Δλ: stepwidth in nm.

λmax: wavelength of the upper boundary in nm.

cmf: color matching function.

ex: extrapolation mode.

λmin, Δλ, λmax

For obvious reasons, λmin, Δλ, and λmax ideally correspond with the parameters given by cmf.

cmf

The field cmf defines which color matching function type is used for all spectral calculations where CMFs are relevant, for example in color deficiency simulations, determination of fundamental metamers, or color calculations.

ex

The field ex determines the mode used to extrapolate spectra beyond the boundaries of their defined range, if necessary.

The following extrapolation modes are available in SpectralVis:

:none, :zero

Spectral data outside the defined range are set to zero.

:boundary

Spectral data outside the defined range are set to the values of the respective lower and upper boundaries.

:linear

Spectral data outside the defined range are linearly extrapolated.

:parabolic

Spectral data outside the defined range are extrapolated by parabolas.

setting up a spectral environment and changing SPECENV fields

set_specenv(λmin = 390.0, Δλ = 1.0, λmax = 830.0, cmf = cmfunc(:cie12_10), ex = :linear)

SpectralVis.set_specenv — Function

set_specenv(λmin=390.0,Δλ=1.0,λmax=830.0,extrapolate= :zero)

Initializes the spectral environment to standard values:

λmin = 380.0 nm λmax = 830.0 nm Δλ = 1.0 nm cmf = cmfunc(:cie12_10) ex = :linear

set_extrap(extrapolation::Symbol)

SpectralVis.set_extrap — Function

set_extrap(env::SpecEnvironment,extrapolation::Symbol)

Sets the extrapolation mode of the environment env to one of the available extrapolation settings:

:none or :zero: out of range values are zero.

:constant: out of range values are set to λmin and λmax, respectively.

:linear: out of range values are linearly extrapolated through the first/last two data points.

:parabolic: out fo range values are extrapolated with a parabola through the first/last three datapoints. Recommended extrapolation method according to D.L. MacAdams in "Color Measurement", chapter 5.4 "Truncations".

set_extrap(extrapolation::Symbol)

Sets the extrapolation mode of the environment env to one of the available extrapolation settings:

:none or :zero: out of range values are zero.

:constant: out of range values are set to λmin and λmax, respectively.

:linear: out of range values are linearly extrapolated through the first/last two data points.

:parabolic: out fo range values are extrapolated with a parabola through the first/last three datapoints. Recommended extrapolation method according to D.L. MacAdams in "Color Measurement", chapter 5.4 "Truncations".

set_limits(env::SpecEnvironment, λmin::Real,Δλ::Real,λmax::Real)

SpectralVis.set_limits — Function

set_limits(env::SpecEnvironment, λmin::Real,Δλ::Real,λmax::Real)

sets the wavelength limits and resolution of the spectral environment

set_limits(λmin::Real,Δλ::Real,λmax::Real)

sets the wavelength limits and resolution of the spectral environment

set_colormatch(env::SpecEnvironment, colormatch::CMatch)

SpectralVis.set_colormatch — Function

set_colormatch(env::SpecEnvironment, colormatch::CMatch)

Sets the color matching function (CMF) used in a defined spectral environment.

Available CMFs:

:cie31, :cie31_j, :cie31_jv, :cie64, :cie12_2, :cie12_10, :lms06_2, :lms06_10

set_colormatch(colormatch::CMatch)

Sets the color matching function (CMF) used in the spectral environment.

Available CMFs:

:cie31, :cie31_j, :cie31_jv, :cie64, :cie12_2, :cie12_10, :lms06_2, :lms06_10