Spectral generators

SpectralVis offers a variety of useful generators for different kinds of spectra.

Block Spectra

SpectralVis.block_spec — Function

block_spec(λmin, λmax, Δλ, λs, λe, str = 1.0, mode = 0)

Regular block spectrum initialized to 0.0, with blocks λs..λe = 1.0

block_spec(λmin, λmax, Δλ, λs, λe, str = `1.0`, 1)

Inversed block spectrum initialized to 1.0, with blocks λs..λe = 0.0

block_spec(env::SpecEnvironment, λ)

Zero block spectrum with only wavelength λ set to 1.0.

Fast version to create highest purity optimal colors. Linear interpolation for intermediate λ values.

block_spec(λ)

Zero block spectrum with only wavelength λ set to 1.0 Fast version to create highest purity optimal colors.

Spectral limits are taken from the spectral environment.

LED spectrum generator

This generator simulates LED type spectra.

SpectralVis.led_spec — Function

led_spec(λmin, λmax, Δλ, λ0, Δλ1_2)

LED spectrum simulation

λ0: peak wavelength

Δλ1_2: half spectral width, width at 50% peak intensity

Result: LSpec

Source: Yoshi Ohno (2004), Simulation Analysis of White LED Spectra and Color Rendering, Proceedings for CIE Expert Sumposium on LED Light Sources, Tokyo, JA (Accessed May 27, 2021)

CIE Daylight Illuminants

SpectralVis.D_series_generator — Function

D_series_generator(T::Real)

The relative spectral power distribution SPD(λ) of a D series illuminant at CCT T, from 300 to 830 nm in 5 nm steps. By definition the output spectra are normalized in a way that their luminance is always 100.0 at a wavelength of 560 nm.

Allowed CCT range: 4000 K ≤ CCT ≤ 25000 K

SpectralVis.D_series_whitepoint — Function

D_series_whitepoint(env::SpecEnv,T::Real)

produces the white point according to the CIE Daylight Series.

result: ::XYZ

D_series_whitepoint(T::Real)

produces the white point according to the CIE Daylight Series.

result: ::XYZ

SpectralVis.D_series_illuminant — Function

D_series_illuminant(env::SpecEnv, T::Real)

produces CIE Daylight spectrum according to desired CCT.

result: ::LSpec

D_series_illuminant(T::Real)

produces CIE Daylight spectrum according to desired CCT.

result: ::LSpec

Daylight illuminant convenience functions

SpectralVis.D50_illuminant — Function

D50_illuminant()

Result: ::LSpec of D50 illuminant, with a CCT of 5003 K

SpectralVis.D55_illuminant — Function

D55_illuminant()

Result: ::LSpec of D55 illuminant, with a CCT of 5503 K

SpectralVis.D65_illuminant — Function

D65_illuminant()

Result: ::LSpec of D55 illuminant, with a CCT of 6504 K

SpectralVis.D75_illuminant — Function

D75_illuminant()

Result: ::LSpec of D75 illuminant, with a CCT of 7504 K

SpectralVis.D93_illuminant — Function

D93_illuminant()

Result: ::LSpec of D93 illuminant, with a CCT of 9304 K

Missing docstring.

Missing docstring for D_series_proper_illuminant. Check Documenter's build log for details.

SpectralVis.normalized_D_series — Function

normalized_D_series(env::SpecEnv = SPECENV, T::Real)

produces CIE Daylight spectrum according to desired CCT, normalized to l=100 at λ=560 nm.

Examples

julia> normalized_D_series(SPECENV, 6500)
LSpec(Real[390.0, 391.0, 392.0, 393.0, 394.0, 395.0, 396.0, 397.0, 398.0, 399.0  …  821.0, 822.0, 823.0, 824.0, 825.0, 826.0, 827.0, 828.0, 829.0, 830.0], Real[0.5460212796545921, 0.5650827295052294, 0.590288902296674, 0.620011293254124, 0.6526213976027774, 0.6864907105678326, 0.7200645866061678, 0.7520838171013811, 0.7813630526687508, 0.8067169439235554  …  0.578028743266789, 0.5810475635333059, 0.5837744258575818, 0.5863553092107371, 0.5889361925638924, 0.5916338590939442, 0.5944483088008923, 0.5973503458905127, 0.6003107745685814, 0.6033003990408742])

normalized_D_series(9300)
LSpec(Real[390.0, 391.0, 392.0, 393.0, 394.0, 395.0, 396.0, 397.0, 398.0, 399.0  …  821.0, 822.0, 823.0, 824.0, 825.0, 826.0, 827.0, 828.0, 829.0, 830.0], Real[0.9296442162075731, 0.9527892027932645, 0.9854018656135435, 1.0248615127313507, 1.0685474522096268, 1.1138389921113123, 1.158318458802078, 1.2003802518585145, 1.2386217891599418, 1.2716404885856807  …  0.4869584146523383, 0.4894740470571454, 0.4917569270423351, 0.4939234308177161, 0.4960899345930971, 0.49834953933632503, 0.5007022450474, 0.5031247764843602, 0.5055938584052438, 0.5080862155680892])

CIE Illuminant Type A

SpectralVis.illuminant_A — Function

illuminant_A(env::SpecEnv = SPECENV)

Produces the spectrum of illuminant A, which represents tungsten-filament lighting, depending on the limits defined by SPECENV. The relative spectral power distribution is equal to a Planckian spectrum at T ≊ 2856 K, which is also its CCT.

CCT according to the International Practical Temperature Scale from 1968.

This is the illuminant that should always be used when incandescent lighting is involved.

Example

julia> illuminant_A()
LSpec(Real[390.0, 391.0, 392.0, 393.0, 394.0, 395.0, 396.0, 397.0, 398.0, 399.0  …  821.0, 822.0, 823.0, 824.0, 825.0, 826.0, 827.0, 828.0, 829.0, 830.0], Real[1.0161486323473047e11, 1.0369227145593657e11, 1.05797774054285e11, 1.0793148447599527e11, 1.1009351284197664e11, 1.1228396592637178e11, 1.1450294713618913e11, 1.1675055649202544e11, 1.1902689060987242e11, 1.2133204268399858e11  …  2.17270715240179e12, 2.1757420157602646e12, 2.178757889232851e12, 2.1817547874185107e12, 2.1847327254181292e12, 2.1876917188295703e12, 2.190631783742765e12, 2.1935529367348203e12, 2.196455194865146e12, 2.199338575670583e12])

CIE Illuminant Type B

SpectralVis.illuminant_B — Function

illuminant_B(env::SpecEnv = SPECENV)

Produces the spectrum of CIE Standard Illuminant B, with a CCT of ≈4874 K. Originally intended to represent sunlight. Now obsolete, but according to Hunt the European brewing industry still uses an approximation of this illuminant.

Example

julia> illuminant_B()
LSpec(Real[390.0, 391.0, 392.0, 393.0, 394.0, 395.0, 396.0, 397.0, 398.0, 399.0  …  821.0, 822.0, 823.0, 824.0, 825.0, 826.0, 827.0, 828.0, 829.0, 830.0], Real[0.313, 0.3223587343061588, 0.3319703445293299, 0.3417825875994216, 0.3517432204463422, 0.3618, 0.3719111024629742, 0.38207638112852754, 0.3923061085625939, 0.4026105573311068  …  7.557164594645486e-6, 9.446455743306857e-6, 7.557164594645484e-6, 3.778582297322749e-6, 1.734723475976807e-20, -2.267149378393646e-6, -3.0228658378581947e-6, -2.6450076081259192e-6, -1.5114329189290965e-6, -2.168404344971009e-21])

CIE Illuminant Type C

SpectralVis.illuminant_C — Function

illuminant_C(env::SpecEnv)

Produces the spectrum of CIE Standard Illuminant C, with a CCT of ≈ 6774 K. Originally intended to represent average daylight.

Example

julia> illuminant_C()
LSpec(Real[390.0, 391.0, 392.0, 393.0, 394.0, 395.0, 396.0, 397.0, 398.0, 399.0  …  821.0, 822.0, 823.0, 824.0, 825.0, 826.0, 827.0, 828.0, 829.0, 830.0], Real[0.474, 0.489271743700384, 0.5046749637120697, 0.5202123118735634, 0.5358864400233714, 0.5517, 0.5676560269088902, 0.583759088923221, 0.6000141374831071, 0.616426124028662  …  0.7121636730114121, 0.7159995912642652, 0.7194836730114121, 0.722791836505706, 0.7261, 0.7295488980965763, 0.7331385307954352, 0.7368337144460056, 0.7405992653977175, 0.7444])

How the CIE Standard Illuminant C is produced:

Standard Source A is combined with a filter made of two 1 cm thick layers, containing two solutions, C₁ and C₂. The solutions are contained in a double cell that is made of a colorless optical glass.

Solution C₁:

3.412 g copper sulphate ($CuSO_4·5H_2O$)

3.412 g mannite [$C_6H_8(OH)_6$]

30.0 ml pyridine ($C_5H_5N$)

1000.0 ml distilled water

Solution C₂:

30.58 g cobalt ammonium sulphate [$CoSO_4·(NH_4)_2SO_4·6H_2O$]

22.52 g copper sulphate ($CuSO_4·5H_2O$)

10.0 ml sulphuric acid (density 1.835 g/ml)

1000.0 ml distilled water

F Series Illuminants

SpectralVis.F_series_illuminant — Function

F_series_illuminant(type::Symbol)

Spectral power distribution of representative fluorescent lamps from 380 to 780 nm in 5 nm steps.

Available types:

:f1 ... :f12

F2, F7, and F11 should be preferred over the other distributions as long as the choice in each of these three groups is not important or critical. The three recommendend illuminants are also marked in the CRI lists below.

For wavelength outside the 380 nm ... 780 nm range the SPDs should be set to zero.

Result: ::LSpec

These are not CIE Standard Illuminants. They were compiled by the CIE as a representative collection of practical fluorescent lamps. The 12 lamps in this collection can be grouped in 3 groups:

normal: two semi-broad band emmisions of antimony and manganese activations in calcium halo-phosphate phosphor.

CRI table

F1: 76

F2: 64

F3: 57

F4: 51

F5: 72

F6: 59

broad-band: enhanced color rendering properties in comparison to the normal group, mostly by using multiple phosphors. Flatter spectral distributions with a wider spectral range.

CRI table

F7: 90

F8: 95

F9: 90

three-band: three narrow-band emissions in the red, green, and blue spectral regions. The narrow-band emissions are produced by ternary compositions of rare-earth phosphors.

CRI table

F10: 81

F11: 83

F12: 83

Examples

julia> F_series_illuminant(:f11)
LSpec(Real[390.0, 391.0, 392.0, 393.0, 394.0, 395.0, 396.0, 397.0, 398.0, 399.0  …  821.0, 822.0, 823.0, 824.0, 825.0, 826.0, 827.0, 828.0, 829.0, 830.0], Real[0.46, 0.442, 0.424, 0.406, 0.388, 0.37, 0.554, 0.738, 0.922, 1.1059999999999999  …  0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0])

julia> F_series_illuminant(:f0)
ERROR: DomainError with f0:
F-series illuminant type does not exist.

F_series_illuminant(n::Int)

Return F series illuminant Fn for n from 1 to 12.

Examples

julia> F_series_illuminant(1)
LSpec(Real[390.0, 391.0, 392.0, 393.0, 394.0, 395.0, 396.0, 397.0, 398.0, 399.0  …  821.0, 822.0, 823.0, 824.0, 825.0, 826.0, 827.0, 828.0, 829.0, 830.0], Real[2.94, 3.046, 3.152, 3.258, 3.3640000000000003, 3.47, 3.81, 4.15, 4.49, 4.83  …  0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0])

julia> F_series_illuminant(13)
ERROR: DomainError with 13:
F series illuminant outside the range 1 ≤ n ≤ 12.

Gas Discharge Illuminant Spectra

SpectralVis.gas_discharge_illuminant — Function

gas_discharge_illuminant(type::Symbol)

SPDs of common gas discharge lamps.

available types:

:lps: Low pressure sodium

:hps: High pressure sodium

:mb: High pressure mercury

:mbf: High pressure mercury with red-emitting phosphor coating, better CRI than MB

:mbtf: High pressure mercury with red-emitting phopsphor coating and tungsten filament ballast, better CRI

:hmi: High pressure mercury with metal halides. Medium arc, iodides, typical stadium lighting, was used in television to supplement daylight

:xenon: Xenon, used in film projectors in cinemas, floodlighting, in lighthouses, for accelerated fading tests, flash photography, lighting fluorescent materials

Example

julia> gas_discharge_illuminant(:lps)
LSpec(Real[390.0, 391.0, 392.0, 393.0, 394.0, 395.0, 396.0, 397.0, 398.0, 399.0  …  821.0, 822.0, 823.0, 824.0, 825.0, 826.0, 827.0, 828.0, 829.0, 830.0], Real[0.0, 0.02, 0.04, 0.06, 0.08, 0.1, 0.08, 0.060000000000000005, 0.04000000000000001, 0.020000000000000004  …  0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0])

julia> gas_discharge_illuminant(:xenon)
LSpec(Real[390.0, 391.0, 392.0, 393.0, 394.0, 395.0, 396.0, 397.0, 398.0, 399.0  …  821.0, 822.0, 823.0, 824.0, 825.0, 826.0, 827.0, 828.0, 829.0, 830.0], Real[96.33, 97.176, 98.022, 98.868, 99.714, 100.56, 101.01, 101.46000000000001, 101.91, 102.36  …  0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0])

gas_discharge_illuminant(n::Int)

Returns gas discharge type illuminant of index n for n from 1 to 7.

Result: ::LSpec

Blackbody Illuminant Spectra

SpectralVis.blackbody_illuminant — Function

blackbody_illuminant(env::SpecEnv,T::Real)

produces blackbody spectrum according to desired blackbody temperature in K.

Examples

SpectralVis.blackbody_whitepoint — Function

blackbody_whitepoint(env::SpecEnv, T::Real)

produces the white point according to desired blackbody temperature in K.

Examples

Sinusoidal Spectra

SpectralVis.sinusoidal_spd — Function

sinusoidal_spd(freq::Real, modfac::Real = 1.0, ampl::Real = 1.0, phase::Real, r = (SPECENV.λmin:SPECENV.Δλ:SPECENV.λmax))

Generates a sinusoidal spectral power distribution. Useful for computational gamut and spectral calculations.